requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==2.0.1
future==0.16.0
idna==2.6
line-bot-sdk==1.5.0
pymongo==3.6.0
pytz==2017.3
requests==2.18.4
urllib3==1.22
